requirements.txt
absl_py==2.1.0
chex==0.1.86
dm_env==1.6
dm_tree==0.1.8
dmlab2d==1.0.0
fire==0.5.0
gymnasium==0.29.1
imageio==2.33.1
immutabledict==4.2.0
joypy==0.2.6
matplotlib==3.8.2
ml_collections==0.1.1
networkx==3.1
nltk==3.8.1
numpy==1.23.5
opencv_python==4.6.0.66
pandas==2.2.2
pygame==2.5.2
PyYAML==6.0.1
PyYAML==6.0.2
ray==2.10.0
reactivex==4.0.4
rich==13.8.0
setuptools==59.6.0
skimage==0.0
stable_baselines3==2.3.2
supersuit==3.9.3
tensorflow==2.12.0
torch==2.4.0
tqdm==4.66.1
